requirements.txt
python-telegram-bot>=20,<21
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
youtube-search-python>=1.6.6
requests>=2.31.0
